started DHOV

+++ script/eval.py
@@ -3,6 +3,10 @@ import matplotlib.pyplot as plt
 import numpy as np
 from scipy.spatial import ConvexHull
 
+from script import dataInit
+from script.dataInit import Rhombus
+
+
 class Plots_for():
     # matplotlib.use('TkAgg')
 
@@ -39,7 +43,7 @@ class Plots_for():
             self.adversarial = adversarial
 
         if adversarial_values is not None:
-            self.adversarial_values = adversarial_values
+            self.adversarial_values = adversarial_values.to(torch.float64)
 
 
 
@@ -53,6 +57,7 @@ class Plots_for():
 
         def test_contained(x):
             bool_val = False
+            #self.model.double()
             if torch.is_tensor(self.c):
                 bool_val = self.model(x).less_equal(self.c)
                 for val in torch.flatten(bool_val):
@@ -126,7 +131,7 @@ class Plots_for():
         y = np.linspace(*self.y_range, 500)
         xx, yy = np.meshgrid(x, y)
         x_in = torch.tensor(np.c_[xx.ravel(), yy.ravel()], dtype=torch.float64)
-
+        #self.model.double()
         y_pred = self.model(x_in)
         y_pred = np.round(y_pred.detach().numpy(), decimals=5).reshape(xx.shape)
 
@@ -162,6 +167,7 @@ class Plots_for():
     def plt_adversarial_dotted(self):
         if self.adversarial is None or self.adversarial_values is None:
             return
+        self.adversarial.double()
 
         pred_x_arr = []
         pred_y_arr = []
@@ -169,6 +175,7 @@ class Plots_for():
         ax = plt.axes()
         ax.set_xlim([self.x_range[0], self.x_range[1]])
         ax.set_ylim([self.y_range[0], self.y_range[1]])
+
 
 
         for x in self.adversarial_values:
